+++ stories/02-Design System/02-Layout/Grid/examples/index.tsx
@@ -1,27 +1,39 @@
 import Basic from './basic';
-import Row from './row';
-import Col from './col';
-import Grid from './grid';
-import GridMenu from './gridMenu';
+import Playground from './playground';
+import Xs from './xs';
+import Sm from './sm';
+import Md from './md';
+import Lg from './lg';
+import Xl from './xl';
+import Xxl from './xxl';
 
 import BasicSource from './basic?raw';
-import RowSource from './row?raw';
-import ColSource from './col?raw';
-import GridSource from './grid?raw';
-import GridMenuSource from './gridMenu?raw';
+import PlaygroundSource from './playground?raw';
+import XsSource from './xs?raw';
+import SmSource from './sm?raw';
+import MdSource from './md?raw';
+import LgSource from './lg?raw';
+import XlSource from './xl?raw';
+import XxlSource from './xxl?raw';
 
 export const Sources = {
     Basic: BasicSource,
-    Row: RowSource,
-    Col: ColSource,
-    Grid: GridSource,
-    GridMenu: GridMenuSource
+    Playground: PlaygroundSource,
+    Xs: XsSource,
+    Sm: SmSource,
+    Md: MdSource,
+    Lg: LgSource,
+    Xl: XlSource,
+    Xxl: XxlSource
 };
 
 export default {
     Basic,
-    Row,
-    Col,
-    Grid,
-    GridMenu
+    Playground,
+    Xs,
+    Sm,
+    Md,
+    Lg,
+    Xl,
+    Xxl
 };